+++ TodoApi/Repository/IRepository.cs
@@ -10,7 +10,7 @@ namespace TodoApi.Repository
     {
         Task<List<T>> GetAsync(params Expression<Func<T, object>>[] includePredicates);
         Task<List<T>> GetAsync(Expression<Func<T, bool>> wherePredicate, params Expression<Func<T, object>>[] includePredicates);
-        Task<T> GetAsync(long id, params Expression<Func<T, object>>[] includePredicates);
+        Task<T> GetAsync(int id, params Expression<Func<T, object>>[] includePredicates);
         Task AddAsync(T entity);
         void Update(T entity);
         void Delete(T entity);
